Fix Bug Cannot Delete FileOnStudentAssignment that is type link-url (https://github.com/tatuga-camp/client-main-tatugaschool/issues/132)

+++ src/file-on-student-assignment/file-on-student-assignment.repository.ts
@@ -146,7 +146,10 @@ export class FileOnStudentAssignmentRepository
           },
         });
 
-      if (fileOnStudentAssignment.contentType === 'FILE') {
+      if (
+        fileOnStudentAssignment.contentType === 'FILE' &&
+        fileOnStudentAssignment.type !== 'link-url'
+      ) {
         await this.googleStorageService.DeleteFileOnStorage({
           fileName: fileOnStudentAssignment.body,
         });
